Emergency's web server/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
psycopg2==2.8.6
paho-mqtt==1.5.1